add sub-tiles sampling

+++ cluster_analysis/utils.py
@@ -2,7 +2,7 @@ import numpy as np
 
 def get_flat(array):
     _, h, w = array.shape
-    data = np.empty((h * w, len(array)), dtype=np.int16)
+    data = np.empty((h * w, len(array)))
     for i in range(len(array)):
         data[:, i] = array[i, :, :].flatten()
